% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_landmark_dataset.R
\name{create_landmark_dataset}
\alias{create_landmark_dataset}
\title{Create landmark dataset}
\usage{
create_landmark_dataset(
  data,
  x_L,
  x_hor,
  patient_id,
  assessment_time,
  event_status,
  event_time
)
}
\arguments{
\item{data}{Data frame with repeated measurements in long format; each row corresponds to an assessment entry}

\item{x_L}{Integer specifying the landmark time}

\item{x_hor}{Integer specifying the horizon time}

\item{patient_id}{Character string specifying the column name in \code{data} which contains the individual identifiers}

\item{assessment_time}{Character string specifying the column name in
\code{data} which contains the assessment time for each row}

\item{event_status}{Character string specifying the column name in \code{data}
which contains the event status (where 0=censoring, 1=event of interest, if there are competing events these are labelled
2 or above.)}

\item{event_time}{Character string specifying the column name in \code{data}
which contains the event time}
}
\value{
Data frame \code{data} updated to contain only entries of
individuals at risk at age \code{x_L}
}
\description{
Selects the rows of individuals at risk at age \code{x_L},
i.e individuals that started follow-up before
(or at) landmark time \code{x_L}, and ended follow-up after
(but not at) landmark time \code{x_L}. These entries comprise
the dataset for landmark time \code{x_L} in the landmark model.
This function also censors individuals with event times over \code{x_hor},
according to the landmarking model.
}
\details{
An individual enters the risk set at their first assessment and exits the risk set
after their first event (this may be censoring, event of interest or a competing event).
}
\examples{
data(data_repeat_outcomes)
data_repeat_outcomes <- create_landmark_dataset(data=data_repeat_outcomes,
  x_L=60,
  x_hor=65,
  assessment_time="response_time_sbp_stnd",
  patient_id="id",
  event_time="event_time",
  event_status="event_status")

}
\author{
Isobel Barrott \email{isobel.barrott@gmail.com}
}
