% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_landmark_cv}
\alias{data_landmark_cv}
\title{Simulated data for use in a landmarking analysis}
\format{
A dataset with 22427 rows and 18 columns.
}
\usage{
data_landmark_cv
}
\description{
A combination of the \code{data_repeat} and \code{data_outcomes} data, modified to perform a landmark analysis. See the vignette associated with the \code{Landmarking}
package (\code{browseVignettes("Landmarking")}) for more information about how this dataset was formed.
}
\keyword{datasets}
