% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_landmark_LOCF}
\alias{data_landmark_LOCF}
\title{Simulated data for use in a landmarking analysis}
\format{
A dataset with 6921 rows and 12 columns.
}
\usage{
data_landmark_LOCF
}
\description{
This dataset includes risk predictions for a CVD event. See the vignette associated with the \code{Landmarking}
package (\code{browseVignettes("Landmarking")}) for more information about how this dataset was formed.
}
\keyword{datasets}
