% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_repeat}
\alias{data_repeat}
\title{Simulated repeat measurement data}
\format{
A dataset with 9048 rows and 11 columns:
\describe{
\item{id}{Patient ID}
\item{smoking}{Smoking status, 0 indicates the patient has never smoked, 1 indicates the patient has quit smoking, and 2 indicates the patient is a current smoker}
\item{diabetes}{Diabetes status, 0 indicates the patient is not diagnosed with diabetes, and 1 indicates the patient is diagnosed with diabetes}
\item{ethnicity}{Ethnicity, one of five ethnicities}
\item{deprivation}{Deprivation score, quintiles}
\item{index}{An index indicating assessment number for a patient}
\item{sbp_stnd}{Standardised systolic blood pressure}
\item{tchdl_stnd}{Standardised total cholesterol to HDL ratio}
\item{end_of_study_age}{Age that the patient left the study, either the age at event (CVD or death) or age at end of study (1st Jan 2010)}
\item{response_time_tchdl_stnd}{Age that total cholesterol to HDL ratio was recorded}
\item{response_time_sbp_stnd}{Age that systolic blood pressure and all other measurements were recorded}
}
}
\usage{
data_repeat
}
\description{
A simulated dataset formed using an LME model, whose parameters were based on CVD risk assessments
recorded at primary care practices in New Zealand. The dataset contains repeat measurements data from 3000 patients.
}
\keyword{datasets}
