% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_landmark_model.R
\name{fit_LME_landmark_model}
\alias{fit_LME_landmark_model}
\title{Fit a landmarking model using a linear mixed effects (LME) model for the longitudinal submodel}
\usage{
fit_LME_landmark_model(
  data,
  x_L,
  x_hor,
  fixed_effects,
  random_effects,
  fixed_effects_time,
  random_effects_time,
  random_intercept = TRUE,
  random_slope = TRUE,
  standardise_time = FALSE,
  cv_name,
  patient_id,
  lme_control = nlme::lmeControl(),
  event_time,
  event_status,
  cr_method
)
}
\arguments{
\item{data}{Data frame containing longitudinal (repeat measurements) data and time to event data, there may be more than one row per individual}

\item{x_L}{Integer specifying the landmark time}

\item{x_hor}{Integer specifying the horizon time}

\item{fixed_effects}{Vector of character strings specifying the column names in \code{data}
which correspond to the fixed effects}

\item{random_effects}{Vector of character strings specifying the column names in \code{data}
which correspond to the random effects}

\item{fixed_effects_time}{Character string specifying the column name in \code{data}
which contains the time at which the fixed effects were recorded}

\item{random_effects_time}{Vector of character strings specifying the column names in \code{data}
which contain the times at which repeat measures were recorded. This should either be length 1 or
the same length as \code{random_effects}. In the latter case the order of elements must correspond to the
order of elements in \code{random_effects}.}

\item{random_intercept}{Boolean indicating whether to include a random intercept in the LME model}

\item{random_slope}{Boolean indicating whether to include a random slope in the LME model}

\item{standardise_time}{Boolean indicating whether to standarise the \code{covariates_time} (see Details section for more information)}

\item{cv_name}{Character string specifying the column name in \code{data}
which contains the cross-validation fold. Leaving this parameter empty (default) means that
no cross-validation will be performed.}

\item{patient_id}{Character string specifying the column name in \code{data} which contains the individual identifiers}

\item{lme_control}{Object create using nlme::lmeControl(), which contains the \code{control} argument in the \code{lme}
function}

\item{event_time}{Character string specifying the column name in \code{data}
which contains the event time}

\item{event_status}{Character string specifying the column name in \code{data}
which contains the event status (0=censoring, 1=event of interest, 2+=competing
events)}

\item{cr_method}{Character string specifying which competing risks model to
use. Three options: the standard Cox model i.e. no competing risks (\code{"standard_cox"}),
the cause-specific regression model (\code{"cause_specific"}), or the Fine Gray
regression model (\code{"fine_gray"})}
}
\value{
List containing \code{data}, \code{model_longitudinal}, and \code{model_survival}.  The
data frame \code{data} contains one row for each individual and includes the \code{covariates}
that are calculated using the LME model (more information in Details section).

\code{model_longitudinal} contains the output from
the \code{lme} function from package \code{nlme}. For a model using cross-validation,
\code{model_longitudinal} will contain a list of outputs with each
element in the list corresponds to a different cross-validation fold.

\code{model_survival} contains the outputs from
the \code{coxph} function from package \code{survival}. This output contains the
estimated parameters for the survival submodel. For a model using cross-validation,
\code{model_survival} will contain a list of outputs with each
element in the list corresponds to a different cross-validation fold.
}
\description{
This function performs the two-stage landmarking analysis. In the first stage, the longitudinal submodel is fitted using the LME model and in the
second stage the survival submodel is fitted.
}
\details{
dfsdfdsfsd
}
\examples{
\dontrun{2^3}
}
\author{
Isobel Barrott \email{isobel.barrott@gmail.com}
}
