% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_LME_landmark_model.R
\name{fit_LME_longitudinal_model}
\alias{fit_LME_longitudinal_model}
\title{Fit a landmarking model using a linear mixed effects (LME) model for the longitudinal submodel}
\usage{
fit_LME_longitudinal_model(
  data_long,
  x_L,
  fixed_effects,
  random_effects,
  fixed_effects_time,
  random_effects_time,
  standardise_time = FALSE,
  random_slope = TRUE,
  cv_name = NA,
  patient_id,
  lme_control = nlme::lmeControl()
)
}
\arguments{
\item{data_long}{Data frame containing repeat measurement data and time-to-event data in long format.}

\item{x_L}{Numeric specifying the landmark time(s)}

\item{fixed_effects}{Vector of character strings specifying the column names in \code{data}
which correspond to the fixed effects}

\item{random_effects}{Vector of character strings specifying the column names in \code{data}
which correspond to the random effects}

\item{fixed_effects_time}{Character string specifying the column name in \code{data}
which contains the time at which the fixed effects were recorded}

\item{random_effects_time}{Vector of character strings specifying the column names in \code{data}
which contain the times at which repeat measures were recorded. This should either be length 1 or
the same length as \code{random_effects}. In the latter case the order of elements must correspond to the
order of elements in \code{random_effects}.}

\item{standardise_time}{Boolean indicating whether to standarise the time variable by subtracting the mean
and dividing by the standard deviation (see Details section for more information)}

\item{random_slope}{Boolean indicating whether to include a random slope in the LME model}

\item{cv_name}{Character string specifying the column name in \code{data_long} that indicates cross-validation fold}

\item{patient_id}{Character string specifying the column name in \code{data_long}/\code{data} which contains the individual identifiers}

\item{lme_control}{Object created using \code{nlme::lmeControl()}, which will be passed to the \code{control} argument of the \code{lme} function}
}
\value{
List containing \code{data_longitudinal}, \code{model_longitudinal}, \code{model_LME}, and \code{model_LME_standardise_time}.

\code{data_longitudinal} has one row for each individual in \code{data_long} and
contains the predicted value at the landmark time \code{x_L}of the \code{fixed_effects} using the LOCF model and
\code{random_effects} using the LME model.

\code{model_longitudinal} indicates that the longitudinal submodel is LME.

\code{model_LME} contains the output from
the \code{lme} function from package \code{nlme}. For a model using cross-validation,
\code{model_LME} contains a list of outputs with each
element in the list corresponds to a different cross-validation fold.

\code{model_LME_standardise_time} contains a list of two objects \code{mean_response_time} and \code{sd_response_time} if the parameter \code{standardise_time=TRUE} is used. This
is the mean and standard deviation use to normalise times when fitting the LME model.
}
\description{
This function is a helper function for \code{fit_LME_landmark_model}.
}
