% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_landmark_model.R
\name{fit_LOCF_landmark_model}
\alias{fit_LOCF_landmark_model}
\title{Fit a landmarking model using a last observation carried forward (LOCF) model for the longitudinal submodel}
\usage{
fit_LOCF_landmark_model(
  data,
  x_L,
  x_hor,
  covariates,
  covariates_time,
  cv_name,
  patient_id,
  event_time,
  event_status,
  survival_submodel = c("standard_cox", "cause_specific", "fine_gray")
)
}
\arguments{
\item{data}{Data frame containing longitudinal (repeat measurements) data and time to event data, there may be more than one row per individual}

\item{x_L}{Integer specifying the landmark time}

\item{x_hor}{Integer specifying the horizon time}

\item{covariates}{Vector of character strings specifying the column names in \code{data}
which correspond to the covariates}

\item{covariates_time}{Vector of character strings specifying the column names in \code{data}
which contain the times at which \code{covariates} were recorded. This should either be length 1 or
the same length as \code{covariates}. In the latter case the order of elements must correspond to the
order of elements in \code{covariates}.}

\item{cv_name}{Character string specifying the column name in \code{data}
which contains the cross-validation fold. Leaving this parameter empty (default) means that
no cross-validation will be performed.}

\item{patient_id}{Character string specifying the column name in \code{data} which contains the individual identifiers}

\item{event_time}{Character string specifying the column name in \code{data}
which contains the event time}

\item{event_status}{Character string specifying the column name in \code{data}
which contains the event status (where 0=censoring, 1=event of interest, if there are competing events these are labelled
2 or above.)}

\item{survival_submodel}{Character string specifying which survival submodel to
use. Three options: the standard Cox model i.e. no competing risks (\code{"standard_cox"}),
the cause-specific regression model (\code{"cause_specific"}), or the Fine Gray
regression model (\code{"fine_gray"})}
}
\value{
List containing \code{data}, \code{model_longitudinal}, and \code{model_survival}.

\code{data} is a data frame that includes the \code{covariates} values
calculated using the LOCF model (see Details section) and the predicted
probability that the event of interest has occured by time \code{x_L}, labelled as "event_prediction".
There is one row for each individual.

\code{model_longitudinal} indicates that the longitudinal submodel is LOCF.

\code{model_survival} contains the outputs from
the survival submodel functions, including the estimated parameters of the model.
For a model using cross-validation,
\code{model_survival} will contain a list of outputs with each
element in the list corresponding to a different cross-validation fold.
}
\description{
This function performs the two-stage landmarking analysis. In the first stage, the longitudinal submodel is fitted using the LOCF model and in the
second stage the survival submodel is fitted.
}
\details{
There are two parts to fitting the landmark model: the longitudinal submodel and the survival submodel.

For the longitudinal model, this function uses the most recent values of the covariates at the landmark
age \code{x_L}. This is the LOCF model.

For the survival submodel, there are three choices of model:
\itemize{
\item the standard Cox model, this is a wrapper function for \code{coxph} from the package \code{survival}
\item the cause-specific model, this is a wrapper function for \code{CSC} from package \code{riskRegression}
\item the Fine Gray model, this is a wrapper function for \code{FGR} from package \code{riskRegression}
}

The latter two models estimate the probability of the event of interest in the presence of competing events.
}
\examples{
data(data_landmark_cv)
data_model_landmark_LOCF<-fit_LOCF_landmark_model(data=data_landmark_cv,
  x_L=60,
  x_hor=65,
  covariates=c("ethnicity","smoking","diabetes","sbp_stnd","tchdl_stnd"),
  covariates_time=c(rep("response_time_sbp_stnd",4),"response_time_tchdl_stnd"),
  cv_name="cross_validation_number",
  patient_id="id",
  event_time="event_time",
  event_status="event_status",
  survival_submodel = "cause_specific")
}
\author{
Isobel Barrott \email{isobel.barrott@gmail.com}
}
