% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_ids_with_LOCF.R
\name{return_ids_with_LOCF}
\alias{return_ids_with_LOCF}
\title{Select individuals in a dataset with a last observation carried forward (LOCF) at a landmark time}
\usage{
return_ids_with_LOCF(data, patient_id, x_L, covariates, covariates_time)
}
\arguments{
\item{data}{Data frame with repeated measurements in long format; each row corresponds to an assessment entry}

\item{patient_id}{Character string specifying the column name in \code{data} which contains the individual identifiers}

\item{x_L}{Numeric specifying the landmark time}

\item{covariates}{Vector of character strings specifying the column names in \code{data}
which correspond to the covariates}

\item{covariates_time}{Vector of character strings specifying the column names in \code{data}
which contain the times at which \code{covariates} were recorded. This should either be length 1 or
the same length as \code{covariates}. In the latter case the order of elements must correspond to the
order of elements in \code{covariates}.}
}
\value{
Data frame \code{data} updated to contain only rows of individuals with a LOCF at age \code{x_L}, other rows are removed
}
\description{
To fit the LOCF model, all individuals must have at least one
non-\code{NA} entry by age \code{x_L} for all covariates. This function selects these individuals and removes the other rows.
}
\details{
Individuals have a LOCF if there is a non-\code{NA} entry for each of the covariates in
\code{covariates} up until (and including) time \code{x_L}.
}
\examples{
data(data_repeat_outcomes)
data_repeat_outcomes<-return_ids_with_LOCF(data=data_repeat_outcomes,
  patient_id="id",
  covariates=c("ethnicity","smoking","diabetes","sbp_stnd","tchdl_stnd"),
  covariates_time=c(rep("response_time_sbp_stnd",4),"response_time_tchdl_stnd"),
  x_L=60)
}
\author{
Isobel Barrott \email{isobel.barrott@gmail.com}
}
