% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_repeat_outcomes}
\alias{data_repeat_outcomes}
\title{Simulated CVD event data}
\format{
An object of class \code{data.frame} with 110021 rows and 17 columns.
}
\usage{
data_repeat_outcomes
}
\description{
A dataset containing CVD event data for 4100 patients at risk at landmark age 60. This is simulated data formed using an cause-specific survival model.
}
\details{
A dataset with 4100 rows and 3 columns:
\describe{
\item{id}{Patient ID}
\item{event_status}{Event status, 0 indicates censoring, 1 indicates CVD event, 2 indicates death from other causes}
\item{event_time}{Event time}
}
}
\keyword{datasets}
