% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_LME_landmark.R
\name{fit_LME_landmark}
\alias{fit_LME_landmark}
\title{Fit a landmarking model using a linear mixed effects (LME) model for the longitudinal submodel}
\usage{
fit_LME_landmark(
  data_long,
  x_L,
  x_hor,
  fixed_effects,
  random_effects,
  fixed_effects_time,
  random_effects_time,
  random_slope = TRUE,
  standardise_time = FALSE,
  individual_id,
  start_study_time,
  end_study_time,
  k,
  cross_validation_df,
  lme_control = nlme::lmeControl(),
  event_time,
  event_status,
  survival_submodel = c("standard_cox", "cause_specific", "fine_gray"),
  b
)
}
\arguments{
\item{data_long}{Data frame or list of data frames each corresponding to a landmark age \code{x_L} (each element of the list must be named the value of \code{x_L} it corresponds to)
Each data frame contains repeat measurements data and time-to-event data in long format.}

\item{x_L}{Numeric specifying the landmark time(s)}

\item{x_hor}{Numeric specifying the horizon time(s)}

\item{fixed_effects}{Vector of character strings specifying the column names in \code{data}
which correspond to the fixed effects}

\item{random_effects}{Vector of character strings specifying the column names in \code{data}
which correspond to the random effects}

\item{fixed_effects_time}{Character string specifying the column name in \code{data}
which contains the time at which the fixed effects were recorded}

\item{random_effects_time}{Vector of character strings specifying the column names in \code{data}
which contain the times at which repeat measures were recorded. This should either be length 1 or
the same length as \code{random_effects}. In the latter case the order of elements must correspond to the
order of elements in \code{random_effects}.}

\item{random_slope}{Boolean indicating whether to include the estimated value of the random slope in the LME model}

\item{standardise_time}{Boolean indicating whether to standardise the time variable by subtracting the mean
and dividing by the standard deviation (see Details section for more information)}

\item{individual_id}{Character string specifying the column name in \code{data_long} which contains the individual identifiers}

\item{start_study_time}{Character string specifying the column name in \code{data} which indicates the time that an individual enters the study}

\item{end_study_time}{Character string specifying the column name in \code{data} which indicates the time that an individual leaves the study}

\item{k}{Integer specifying the number of folds for cross-validation}

\item{cross_validation_df}{List of data frames containing the cross-validation fold each individual is assigned to. Each data frame in the list should be
named according to the landmark time \code{x_L} that they correspond. Each data frame should contain the columns \code{individual_id} and a column \code{cross_validation_number}
which contains the cross-validation fold of the individual. An alternative to setting parameter \code{k} for performing cross-validation;
if both are missing no cross-validation is used.}

\item{lme_control}{Object created using \code{nlme::lmeControl()}, which will be passed to the \code{control} argument of the \code{lme}
function}

\item{event_time}{Character string specifying the column name in \code{data_long}
which contains the event time}

\item{event_status}{Character string specifying the column name in \code{data_long}
which contains the event status (where 0=censoring, 1=event of interest, if there are competing events these are labelled
2 or above). Events at time \code{x_hor} should be labelled censored.}

\item{survival_submodel}{Character string specifying which survival submodel to
use. Three options: the standard Cox model i.e. no competing risks (\code{"standard_cox"}),
the cause-specific regression model (\code{"cause_specific"}), or the Fine Gray
regression model (\code{"fine_gray"})}

\item{b}{Integer specifying the number of bootstrap samples to take when calcluating standard error of c-index and Brier score}
}
\value{
List containing \code{data}, \code{model_longitudinal}, \code{model_LME}, \code{model_LME_standardise_time}, \code{model_survival}, and \code{prediction_error}.

\code{data} is a data frame that includes the LOCF values of the \code{fixed_effects} and estimates
of the \code{random_effects} predicted from the LME model (see Details section). It also includes the predicted
probability that the event of interest has occured by time \code{x_L}, labelled as \code{"event_prediction"}.
There is one row for each individual.

\code{model_longitudinal} indicates that the longitudinal submodel is LME.

\code{model_LME} contains the output from
the \code{lme} function from package \code{nlme}. For a model using cross-validation,
\code{model_LME} contains a list of outputs with each
element in the list corresponds to a different cross-validation fold.
\code{prediction_error} contains a list indicating the c-index and Brier score at time \code{x_hor} and their standard errors if parameter \code{b} is used.
For more information on how the prediction error is calculated
please see \code{?get_model_assessment} which is the function used to do this within \code{fit_LME_landmark}.

\code{model_LME_standardise_time} contains a list of two objects \code{mean_response_time} and \code{sd_response_time} if the parameter \code{standardise_time=TRUE} is used. This
is the mean and standard deviation use to normalise times when fitting the LME model.

\code{model_survival} contains the outputs from
the survival submodel functions, including the estimated parameters of the model. For a model using cross-validation,
\code{model_survival} will contain a list of outputs with each
element in the list corresponding to a different cross-validation fold.
\code{model_survival} contains the outputs from the function used to fit the survival submodel, including the estimated parameters of the model.
For a model using cross-validation, \code{model_survival} contains a list of outputs with each
element in the list corresponding to a different cross-validation fold. For more information on how the survival model is fitted
please see \code{?fit_survival_model} which is the function used to do this within \code{fit_LME_landmark}.

\code{prediction_error} contains a list indicating the c-index and Brier score at time \code{x_hor} and their standard errors if parameter \code{b} is used.
}
\description{
This function performs the two-stage landmarking analysis. In the first stage the longitudinal submodel is fitted using the LME model and in the
second stage the survival submodel is fitted.
}
\details{
Firstly this function selects the individuals in the risk set at the landmark time \code{x_L}. Specifically, the individuals in the risk set are those that have entered the study before the landmark age
(\code{start_study_time} is less than \code{x_L}) and exited the study on or after the landmark age (\code{end_study_time} is the same as or more than \code{x_L})). If the option to use cross validation
is selected, this function then assigns the individuals in the risk set to cross-validation folds and fits the landmark model using training and test datasets. If cross-validation is not selected then the landmark model is
fit to the entire group of individuals in the risk set. The performance of the model is then assessed on the set of predictions from the entire set of individuals in the risk set
by calculating Brier score and C-index.

There are two parts to fitting the landmark model: using the longitudinal data and using the survival data.
This function is used to fit the LME model to the longitudinal data. This model will be described in more detail.

For an individual \eqn{i}, the LME model can be written as

\deqn{Y_i = X_i \beta + Z_i U_i + \epsilon_i}

where
\itemize{
\item \eqn{Y_i} is the vector of outcomes at different time points for the individual
\item \eqn{X_i} is the matrix of covariates for the fixed effects at these time points
\item \eqn{\beta} is the vector of coefficients for the fixed effects
\item \eqn{Z_i} is the matrix of covariates for the random effects
\item \eqn{U_i} is the matrix of coefficients for the random effects
\item \eqn{\epsilon_i} is the error term, typically from N(0, \eqn{\sigma})
}

By using an LME model to fit repeat measures data we can allow measurements from the same individuals to be
more similar than measurements from different individuals. This is done through the random intercept and/or
random slope.

Extending this model to the case where there are multiple random effects, denoted \eqn{k}, we have

\deqn{Y_{ik} = X_{ik} \beta_k + Z_{ik} U_{ik} + \epsilon_{ik}}

Using this model we can allow a covariance structure within the random effects term \eqn{U_{ik}}, for example a sample from the
multivariate normal (MVN) distribution \eqn{MVN(0,\Sigma_u)}. This covariance structure means the value of one random effects variable informs about the
value of the other random effects variables, leading to more accurate predictions and allowing there to be missing data in the
random effects variables.

The function \code{fit_LME_landmark} uses this covariance structure for the random effects when fitting the LME model.
To fit the LME model the function \code{lme} from the package \code{nlme} is used.
The fixed effects are calculated as the LOCF for the variables \code{fixed_effects} at the landmark age \code{x_L} and the random effects
are those stated in \code{random_effects} and at times \code{random_effects_time}. This model is used to predict the
values of the random effects at the landmark time \code{x_L}.

It is important to distinguish between the validation set and the development set for fitting the LME model. The development set includes
all the repeat measurements (including those after the landmark age \code{x_L}). Conversely, the validation set only includes
the repeat measurements recorded up until and including the landmark age \code{x_L}.

There is an important consideration about fitting the linear mixed effects model. As the variable \code{random_effects_time}
gets further from 0, the random effects coefficients get closer to 0. This causes computational issues
as the elements in the covariance matrix of the random effects, \eqn{\Sigma_u}, are constrained to
be greater than 0. Using parameter \code{standard_time=TRUE} can prevent this issue by standardising the
time variables to ensure that the \code{random_effects_time} values are not too close to 0.

The predictions of the random effects at the landmark age, in addition to the LOCF values for the fixed effects,
are used as the covariates for the survival submodel.
The values of these covariates can be viewed in the data frame \code{data} that are returned by this function.

For the survival submodel, there are three choices of model:
\itemize{
\item the standard Cox model, this is a wrapper function for \code{coxph} from the package \code{survival}
\item the cause-specific model, this is a wrapper function for \code{CSC} from package \code{riskRegression}
\item the Fine Gray model, this is a wrapper function for \code{FGR} from package \code{riskRegression}
}

The latter two models estimate the probability of the event of interest in the presence of competing events.
}
\examples{
\dontrun{
library(Landmarking)
data(data_repeat_outcomes)
data_repeat_outcomes <-
  return_ids_with_LOCF(
    data_long = data_repeat_outcomes,
    individual_id = "id",
    covariates =
      c("ethnicity", "smoking", "diabetes", "sbp_stnd", "tchdl_stnd"),
    covariates_time =
      c(rep("response_time_sbp_stnd", 4), "response_time_tchdl_stnd"),
    x_L = c(60,61)
  )
data_model_landmark_LME <-
  fit_LME_landmark(
    data_long = data_repeat_outcomes,
    x_L = c(60, 61),
    x_hor = c(65, 66),
    k = 10,
    start_study_time = "start_time",
    end_study_time = "event_time",
    fixed_effects = c("ethnicity", "smoking", "diabetes"),
    fixed_effects_time = "response_time_sbp_stnd",
    random_effects = c("sbp_stnd", "tchdl_stnd"),
    random_effects_time = c("response_time_sbp_stnd", "response_time_tchdl_stnd"),
    individual_id = "id",
    standardise_time = TRUE,
    lme_control = nlme::lmeControl(maxIter = 100, msMaxIter = 100),
    event_time = "event_time",
    event_status = "event_status",
    survival_submodel = "cause_specific"
  )
}
}
\author{
Isobel Barrott \email{isobel.barrott@gmail.com}
}
