% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_landmark_model.R
\name{fit_LME_landmark_model}
\alias{fit_LME_landmark_model}
\title{Fit a landmarking model using a linear mixed effects (LME) model for the longitudinal submodel}
\usage{
fit_LME_landmark_model(
  data,
  x_L,
  x_hor,
  fixed_effects,
  random_effects,
  fixed_effects_time,
  random_effects_time,
  random_intercept = TRUE,
  random_slope = TRUE,
  standardise_time = FALSE,
  cv_name,
  patient_id,
  lme_control = nlme::lmeControl(),
  event_time,
  event_status,
  survival_submodel
)
}
\arguments{
\item{data}{Data frame containing repeat measurements data and time-to-event data, there may be more than one row per individual}

\item{x_L}{Integer specifying the landmark time}

\item{x_hor}{Integer specifying the horizon time}

\item{fixed_effects}{Vector of character strings specifying the column names in \code{data}
which correspond to the fixed effects}

\item{random_effects}{Vector of character strings specifying the column names in \code{data}
which correspond to the random effects}

\item{fixed_effects_time}{Character string specifying the column name in \code{data}
which contains the time at which the fixed effects were recorded}

\item{random_effects_time}{Vector of character strings specifying the column names in \code{data}
which contain the times at which repeat measures were recorded. This should either be length 1 or
the same length as \code{random_effects}. In the latter case the order of elements must correspond to the
order of elements in \code{random_effects}.}

\item{random_intercept}{Boolean indicating whether to include a random intercept in the LME model}

\item{random_slope}{Boolean indicating whether to include a random slope in the LME model}

\item{standardise_time}{Boolean indicating whether to standarise the time variable by subtracting the mean
and dividing by the standard deviation (see Details section for more information)}

\item{cv_name}{Character string specifying the column name in \code{data}
which contains the cross-validation fold. Leaving this parameter empty (default) means that
no cross-validation will be performed.}

\item{patient_id}{Character string specifying the column name in \code{data} which contains the individual identifiers}

\item{lme_control}{Object created using \code{nlme::lmeControl()}, which will be passed to the \code{control} argument of the \code{lme}
function}

\item{event_time}{Character string specifying the column name in \code{data}
which contains the event time}

\item{event_status}{Character string specifying the column name in \code{data}
which contains the event status (where 0=censoring, 1=event of interest, if there are competing events these are labelled
2 or above.)}

\item{survival_submodel}{Character string specifying which survival submodel to
use. Three options: the standard Cox model i.e. no competing risks (\code{"standard_cox"}),
the cause-specific regression model (\code{"cause_specific"}), or the Fine Gray
regression model (\code{"fine_gray"})}
}
\value{
List containing \code{data}, \code{model_longitudinal}, and \code{model_survival}.

\code{data} is a data frame that includes the LOCF values of the \code{fixed_effects} and estimates
of the \code{random_effects} predicted from the LME model (see Details section). It also includes the predicted
probability that the event of interest has occured by time \code{x_L}, labelled as "event_prediction".
There is one row for each individual.
\code{model_longitudinal} contains the output from
the \code{lme} function from package \code{nlme}. For a model using cross-validation,
\code{model_longitudinal} will contain a list of outputs with each
element in the list corresponds to a different cross-validation fold.

\code{model_survival} contains the outputs from
the survival submodel functions, including the estimated parameters of the model. For a model using cross-validation,
\code{model_survival} will contain a list of outputs with each
element in the list corresponding to a different cross-validation fold.
}
\description{
This function performs the two-stage landmarking analysis. In the first stage the longitudinal submodel is fitted using the LME model and in the
second stage the survival submodel is fitted.
}
\details{
There are two parts to fitting the landmark model: the longitudinal submodel and the survival submodel.
This function id used to fit the LME model as the longitudinal submodel. This model will be described in more detail.

For an individual \eqn{i}, the LME model can be written as

\deqn{Y_i = X_i \beta + Z_i U_i + \epsilon_i}

where
\itemize{
\item \eqn{Y_i} is the vector of outcomes at different time points for the individual
\item \eqn{X_i} is the matrix of covariates for the fixed effects at these time points
\item \eqn{\beta} is the vector of coefficients for the fixed effects
\item \eqn{Z_i} is the matrix of covariates for the random effects
\item \eqn{U_i} is the matrix of coefficients for the random effects
\item \eqn{\epsilon_i} is the error term, typically from N(0, \eqn{\sigma})
}

By using an LME model to fit repeat measures data we can allow measurements from the same individuals to be
more similar than measurements from different individuals. This is done through the random intercept and/or
random slope.

Extending this model to the case where there are multiple random effects, denoted \eqn{k}, we have

\deqn{Y_{ik} = X_{ik} \beta_k + Z_{ik} U_{ik} + \epsilon_{ik}}

Using this model we can allow a covariance structure within the random effects term \eqn{U_{ik}}, for example a sample from the
multivariate normal (MVN) distribution \eqn{MVN(0,\Sigma_u)}. This covariance structure means the value of one random effects variable informs about the
value of the other random effects variables, leading to more accurate predictions and allowing there to be missing data in the
random effects variables.

The function \code{fit_LME_landmark_model} uses this covariance structure for the random effects when fitting the LME model.
To fit the LME model the function \code{lme} from the package \code{nlme} is used.
The fixed effects are calculated as the LOCF for the variables \code{fixed_effects} at the landmark age \code{x_L} and the random effects
are those stated in \code{random_effects} and at times \code{random_effects_time}. This model is used to predict the
values of the random effects at the landmark time \code{x_L}.

It is important to distinguish between the validation set and the development set for fitting the LME model. The development set includes
all the repeat measurements (including those after the landmark age \code{x_L}). Conversely, the validation set only includes
the repeat measurements recorded up until and including the landmark age \code{x_L}.

There is an important consideration about fitting the linear mixed effects model. As the variable \code{random_effects_time}
gets further from 0, the random effects coefficients get closer to 0. This causes computational issues
as the elements in the covariance matrix of the random effects, \eqn{\Sigma_u}, are constrained to
be greater than 0. Using parameter \code{standard_time=TRUE} can prevent this issue by standardising the
time variables to ensure that the \code{random_effects_time} values are not too close to 0.

The predictions of the random effects at the landmark age, in addition to the LOCF values for the fixed effects,
are used as the covariates for the survival submodel.
The values of these covariates can be viewed in the data frame \code{data} that are returned by this function.

For the survival submodel, there are three choices of model:
\itemize{
\item the standard Cox model, this is a wrapper function for \code{coxph} from the package \code{survival}
\item the cause-specific model, this is a wrapper function for \code{CSC} from package \code{riskRegression}
\item the Fine Gray model, this is a wrapper function for \code{FGR} from package \code{riskRegression}
}

The latter two models estimate the probability of the event of interest in the presence of competing events.
}
\examples{
\dontrun{data(data_landmark_cv)
data_model_landmark_LME<-fit_LME_landmark_model(data=data_landmark_cv,
  x_L=60,
  x_hor=65,
  fixed_effects=c("ethnicity","smoking","diabetes",
  "deprivation","atrial_fibrillation"),
  fixed_effects_time="response_time_sbp_stnd",
  random_effects=c("sbp_stnd","tchdl_stnd"),
  random_effects_time=c("response_time_sbp_stnd","response_time_tchdl_stnd"),
  cv_name="cross_validation_number",
  patient_id="id",
  standardise_time = TRUE,
  lme_control = nlme::lmeControl(maxIter=100,msMaxIter=100),
  event_time="event_time",
  event_status="event_status",
  survival_submodel = "cause_specific")
}
}
\author{
Isobel Barrott \email{isobel.barrott@gmail.com}
}
