% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_landmark_model.R
\name{fit_LME_landmark_model}
\alias{fit_LME_landmark_model}
\title{Fit a landmarking model using a linear mixed effects (LME) model for the longitudinal submodel}
\usage{
fit_LME_landmark_model(
  data,
  x_L,
  x_hor,
  fixed_effects,
  random_effects,
  fixed_effects_time,
  random_effects_time,
  random_intercept = TRUE,
  random_slope = TRUE,
  standardise_time = FALSE,
  cv_name,
  patient_id,
  lme_control = nlme::lmeControl(),
  event_time,
  event_status,
  cr_method
)
}
\arguments{
\item{data}{Data frame containing longitudinal (repeat measurements) data and time to event data, there may be more than one row per individual}

\item{x_L}{Integer specifying the landmark time}

\item{x_hor}{Integer specifying the horizon time}

\item{fixed_effects}{Vector of character strings specifying the column names in \code{data}
which correspond to the fixed effects}

\item{random_effects}{Vector of character strings specifying the column names in \code{data}
which correspond to the random effects}

\item{fixed_effects_time}{Character string specifying the column name in \code{data}
which contains the time at which the fixed effects were recorded}

\item{random_effects_time}{Vector of character strings specifying the column names in \code{data}
which contain the times at which repeat measures were recorded. This should either be length 1 or
the same length as \code{random_effects}. In the latter case the order of elements must correspond to the
order of elements in \code{random_effects}.}

\item{random_intercept}{Boolean indicating whether to include a random intercept in the LME model}

\item{random_slope}{Boolean indicating whether to include a random slope in the LME model}

\item{standardise_time}{Boolean indicating whether to standarise the \code{covariates_time} (see Details section for more information)}

\item{cv_name}{Character string specifying the column name in \code{data}
which contains the cross-validation fold. Leaving this parameter empty (default) means that
no cross-validation will be performed.}

\item{patient_id}{Character string specifying the column name in \code{data} which contains the individual identifiers}

\item{lme_control}{Object create using nlme::lmeControl(), which contains the \code{control} argument in the \code{lme}
function}

\item{event_time}{Character string specifying the column name in \code{data}
which contains the event time}

\item{event_status}{Character string specifying the column name in \code{data}
which contains the event status (0=censoring, 1=event of interest, 2+=competing
events)}

\item{cr_method}{Character string specifying which competing risks model to
use. Three options: the standard Cox model i.e. no competing risks (\code{"standard_cox"}),
the cause-specific regression model (\code{"cause_specific"}), or the Fine Gray
regression model (\code{"fine_gray"})}
}
\value{
List containing \code{data}, \code{model_longitudinal}, and \code{model_survival}.  The
data frame \code{data} contains one row for each individual and includes the \code{covariates}
that are calculated using the LME model (more information in Details section).

\code{model_longitudinal} contains the output from
the \code{lme} function from package \code{nlme}. For a model using cross-validation,
\code{model_longitudinal} will contain a list of outputs with each
element in the list corresponds to a different cross-validation fold.

\code{model_survival} contains the outputs from
the \code{coxph} function from package \code{survival}. This output contains the
estimated parameters for the survival submodel. For a model using cross-validation,
\code{model_survival} will contain a list of outputs with each
element in the list corresponds to a different cross-validation fold.
}
\description{
This function performs the two-stage landmarking analysis. In the first stage, the longitudinal submodel is fitted using the LME model and in the
second stage the survival submodel is fitted.
}
\details{
There are two parts to fitting the landmark model: the longitudinal (LME) submodel and the survival submodel.

Firstly, the longitudinal submodel (LME) is described in more detail.
For an individual \eqn{i}, the LME model can be written as

\deqn{Y_i = X_i \beta + Z_i U_i + \epsilon_i}

where
\itemize{
\item \eqn{Y_i} is the vector of outcomes at different time points for the individual
\item \eqn{X_i} is the matrix of covariates for the fixed effects at these time points
\item \eqn{\beta} is the vector of coefficients for the fixed effects
\item \eqn{Z_i} is the matrix of covariates for the random effects
\item \eqn{U_i} is the matrix of coefficients for the random effects
\item \eqn{\epsilon_i} is the error term, typically from N(0, \eqn{\sigma})
}

By using an LME model to fit repeat measures data we can allow measurements from the same individuals to be
more similar than measurements from different individuals. This is done through the random intercept and/or
random slope.

Extending this model to the case where there are multiple random effects, denoted \eqn{k}

\deqn{Y_{ik} = X_{ik} \beta_k + Z_{ik} U_{ik} + \epsilon_{ik}}

Using this model we can allow a covariance structure within the random effects term \eqn{U_{ik}}, i.e. it follows the
multivariate normal (MVN) distribution \eqn{MVN(0,\Sigma_u)}. This means information from one random effects variable informs about the
value of the other random effects variables, leading to more accurate predictions and allowing there to be missing data in the
random effects variables.

This function uses the LME model with a covariance structure for the random effects that has been described
for the longitudinal submodel. This is fitted using the function \code{lme} from the package \code{nlme}.
The fixed effects are calculated as the LOCF for the variables \code{fixed_effects} at the landmark age \code{x_L} and the random effects
are those stated in \code{random_effects} and times \code{random_effects_time}. This model is used to predict the
random effects at the landmark time \code{x_L}.
#'
It is important to distinguish between the validation set and the development set for fitting the LME. The development set includes
all the repeat measurements (including those after the landmark age \code{x_L}). Conversely, the validation set only includes
the repeat measurements recorded up until and including the landmark age \code{x_L}.

There is an important consideration about fitting the linear mixed effects model. If the variable \code{random_effects_time}
is far from 0 this means that the random effects coefficients can be close to 0. This causes computational issues
as the variance of the random effects is constrained to
be greater than zero. Therefore the parameter \code{standard_time=TRUE} is it standardises the
\code{random_effects_time} by subtracting the mean of this variable and dividing by the standard deviation.

The predictions of the random effects, in addition to the LOCF values for the fixed effects
are the covariates for the survival submodel. These covariates can be viewed in the data frame \code{data} that are returned by this function.

For the survival submodel, there are three choices of model: the standard Cox model, the cause-specific model and the Fine Gray model.
The latter two models estimate the probability of the event of interest in the presence of competing events.

This function uses the function: \code{coxph} from the pacakge \code{survival} to fit the standard Cox model;
the function \code{CSC} from package \code{riskRegression} to fit the cause-specific model; and the function \code{FGR} from package
\code{riskRegression} to fit the Fine Gray model.
}
\examples{
\dontrun{2^3}
}
\author{
Isobel Barrott \email{isobel.barrott@gmail.com}
}
