% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_landmark_model.R
\name{fit_LOCF_landmark_model}
\alias{fit_LOCF_landmark_model}
\title{Fit a landmarking model using a last observation carried forward (LOCF) model for the longitudinal submodel}
\usage{
fit_LOCF_landmark_model(
  data,
  x_L,
  x_hor,
  covariates,
  covariates_time,
  cv_name,
  patient_id,
  event_time,
  event_status,
  cr_method = c("standard_cox", "cause_specific", "fine_gray")
)
}
\arguments{
\item{data}{Data frame containing longitudinal (repeat measurements) data and time to event data, there may be more than one row per individual}

\item{x_L}{Integer specifying the landmark time}

\item{x_hor}{Integer specifying the horizon time}

\item{covariates}{Vector of character strings specifying the column names in \code{data}
which correspond to the covariates}

\item{covariates_time}{Vector of character strings specifying the column names in \code{data}
which contain the times at which repeat measures were recorded. This should either be length 1 or
the same length as \code{covariates}. In the latter case the order of elements must correspond to the
order of elements in \code{covariates}.}

\item{cv_name}{Character string specifying the column name in \code{data}
which contains the cross-validation fold. Leaving this parameter empty (default) means that
no cross-validation will be performed.}

\item{patient_id}{Character string specifying the column name in \code{data} which contains the individual identifiers}

\item{event_time}{Character string specifying the column name in \code{data}
which contains the event time}

\item{event_status}{Character string specifying the column name in \code{data}
which contains the event status (0=censoring, 1=event of interest, 2+=competing
events)}

\item{cr_method}{Character string specifying which competing risks model to
use. Three options: the standard Cox model i.e. no competing risks (\code{"standard_cox"}),
the cause-specific regression model (\code{"cause_specific"}), or the Fine Gray
regression model (\code{"fine_gray"})}
}
\value{
List containing \code{data}, \code{model_longitudinal}, and \code{model_survival}.  The
data frame \code{data} contains one row for each individual and includes the \code{covariates}
that are calculated using the LOCF model (more information in Details section).

\code{model_longitudinal} indicates that the longitudinal submodel is LOCF.

\code{model_survival} contains the outputs from
the \code{coxph} function from package \code{survival}. This output contains the
estimated parameters for the survival submodel. For a model using cross-validation,
\code{model_survival} will contain a list of outputs with each
element in the list corresponds to a different cross-validation fold.
}
\description{
This function performs the two-stage landmarking analysis. In the first stage, the longitudinal submodel is fitted using the LOCF model and in the
second stage the survival submodel is fitted.
}
\details{
dfsdfdsfsd
}
\examples{
\dontrun{data(data_landmark_cv)
data_model_landmark_LME<-fit_LME_landmark_model(data=data_landmark_cv,x_L=60,x_hor=65,
  fixed_effects=c("ethnicity","smoking","diabetes","deprivation","atrial_fibrillation"),
  fixed_effects_time=rep("response_time_sbp_stnd",length(fixed_effects_col)),
  random_effects=random_effects_col,random_effects_time=random_effects_time_col,
  cv_name="cross_validation_number",patient_id=patient_id_col,
  lme_control = nlme::lmeControl(maxIter=100,msMaxIter=100,
  event_time="event_time",event_status="event_status",cr_method = "cause_specific")}
}
\author{
Isobel Barrott \email{isobel.barrott@gmail.com}
}
