% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixoutsamp.R
\name{mixoutsamp}
\alias{mixoutsamp}
\title{Calculate point estimates from a linear mixed effects (LME) model for new data}
\usage{
mixoutsamp(model, newdata)
}
\arguments{
\item{model}{Object of class \code{lme} containing the fitted LME model}

\item{newdata}{Data frame containing data for which to make predictions}
}
\value{
List containing \code{preddata} and \code{random}. Data frame \code{preddata} is a version of \code{newdata} updated to contain columns conrresponding to the fixed effects values (\code{fixed}),
random effects values (\code{random}), and fitted values (\code{fitted}).
Data frame \code{random} contains the values of random effects components for each individual.
}
\description{
This function allows the user to make out-of-sample point estimate predictions using a LME model.
}
\author{
Ruth Keogh
}
